pyproject.toml
[project]
name = "finance-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=9.3.0",
    "plotly>=6.1.2",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.0",
]
